+++ runtime/runtime/shared/Log.h
@@ -14,17 +14,17 @@ auto get_client_logger() -> const std::shared_ptr<spdlog::logger>&;
 
 } // namespace basalt::log
 
-#ifdef BS_BUILD
+#ifdef BASALT_BUILD
 
-#ifdef BS_DEBUG_BUILD
+#ifdef BASALT_DEBUG_BUILD
 #define BASALT_LOG_TRACE(...) \
   ::basalt::log::get_core_logger()->trace(__VA_ARGS__)
 #define BASALT_LOG_DEBUG(...) \
   ::basalt::log::get_core_logger()->debug(__VA_ARGS__)
-#else
+#else // BASALT_DEBUG_BUILD
 #define BASALT_LOG_TRACE(...)
 #define BASALT_LOG_DEBUG(...)
-#endif
+#endif // BASALT_DEBUG_BUILD
 #define BASALT_LOG_INFO(...) ::basalt::log::get_core_logger()->info(__VA_ARGS__)
 #define BASALT_LOG_WARN(...) ::basalt::log::get_core_logger()->warn(__VA_ARGS__)
 #define BASALT_LOG_ERROR(...) \
@@ -32,17 +32,17 @@ auto get_client_logger() -> const std::shared_ptr<spdlog::logger>&;
 #define BASALT_LOG_FATAL(...) \
   ::basalt::log::get_core_logger()->critical(__VA_ARGS__)
 
-#else // !BS_BUILD
+#else // BASALT_BUILD
 
-#ifdef BS_APP_DEBUG_BUILD
+#ifdef BASALT_DEBUG_BUILD
 #define BASALT_LOG_TRACE(...) \
   ::basalt::log::get_client_logger()->trace(__VA_ARGS__)
 #define BASALT_LOG_DEBUG(...) \
   ::basalt::log::get_client_logger()->debug(__VA_ARGS__)
-#else
+#else // BASALT_DEBUG_BUILD
 #define BASALT_LOG_TRACE(...)
 #define BASALT_LOG_DEBUG(...)
-#endif
+#endif // BASALT_DEBUG_BUILD
 #define BASALT_LOG_INFO(...) \
   ::basalt::log::get_client_logger()->info(__VA_ARGS__)
 #define BASALT_LOG_WARN(...) \
@@ -52,6 +52,6 @@ auto get_client_logger() -> const std::shared_ptr<spdlog::logger>&;
 #define BASALT_LOG_FATAL(...) \
   ::basalt::log::get_client_logger()->critical(__VA_ARGS__)
 
-#endif // BS_BUILD
+#endif // BASALT_BUILD
 
 #endif // !BASALT_SHARED_LOG_H